windows-program/requirements.txt
pyautogui
opencv-python
pynput
websocket-client
